requirements.txt
anyio==4.4.0
certifi==2024.6.2
colorama==0.4.6
dirtyjson==1.0.8
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ollama==0.2.1
python-dotenv==1.0.1
simple-term-menu==1.6.4
sniffio==1.3.1
typing_extensions==4.12.1
